Fenced frame: add URN to inner config
In order to have inner configuration files be able to navigate, they
need to store the URN that the fenced frame will navigate to. Right now,
the new property will not be set to anything, which means it will not
be used to navigate. Once fenced frame configurations are properly
being passed from browser to renderer, this will be used and will
need testing.

Change-Id: I50f634ba5bdcceb6dca78268a2a32952cbf4a537
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/4067741
Reviewed-by: Dominic Farolino <dom@chromium.org>
Commit-Queue: Liam Brady <lbrady@google.com>
Cr-Commit-Position: refs/heads/main@{#1078664}
NOKEYCHECK=True
GitOrigin-RevId: 6a993350b8fbb694197c508c42e294f010932039