Restore mojo::ReportBadMessage() for duplicate MediaPlayerIds.
The underlying cause was fixed some time ago, but the bad message
reporting wasn't restored. This restores the reporting and attempts
to make a surrounding areas more readable.

There are new issues with the new document picture and picture
feature that is being exposed to the fuzzers via experimental
web platform features toggle. For now ensure we end up with
proper process crashes when a duplicate ID ends up being used.

Ultimately it looks like the MediaPlayerHost for the original
document is being used instead of the one in the destination
document. It's unclear to me how this is happening and simply
reconnecting the MediaPlayerHost in DidMoveToNewDocument was
insufficient.

R=liberato

Bug: 1376491, 1393590, 1172882, 1392441
Change-Id: I665595f38f98f7449979f0018e4e236956292059
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/4074026
Auto-Submit: Dale Curtis <dalecurtis@chromium.org>
Commit-Queue: Frank Liberato <liberato@chromium.org>
Reviewed-by: Frank Liberato <liberato@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1078668}
NOKEYCHECK=True
GitOrigin-RevId: 147db76c9ae1281c81077a1a19bd298e8076a1c2