[Instrumented libraries] Rebuild
The previous build was not a clean build, and was from before libc
was removed from the instrumented libraries.  Leaving libc in is not
an issue on the Focal bots, but will lead to runtime errors when
trying to run on other systems such as gLinux.  This CL clean builds
the instrumented libraries to remove libc.

R=thestig

Cq-Include-Trybots: luci.chromium.try:linux_chromium_msan_rel_ng
Change-Id: I271c8b72bd65843d87e6f17b6d9138e802947b57
Bug: 1260217
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/4073981
Reviewed-by: Lei Zhang <thestig@chromium.org>
Commit-Queue: Thomas Anderson <thomasanderson@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1078726}
NOKEYCHECK=True
GitOrigin-RevId: a827eb530d32f2554e051da21e5ab4aa15302de0