[BREFs] Access RuntimeFeatureState(Read)Context from the browser
For WebContentsObservers in the browser process to access and alter the
Blink Runtime-Enabled Features before navigation commit, we need to add
the RuntimeFeatureStateContext to the NavigationRequest.

After commit, we need to store the state of these features in a
RuntimeFeatureStateReadContext object associated with the
RenderFrameHostImpl (via a DocumentUserData helper class). This object
will be used as a read-only reference by the browser process.

Finally, we need to add integration tests to ensure that feature values
are accessed, altered, and persisted correctly.

Design Doc: go/brefs

No-Presubmit: true
Bug: 1377000
Change-Id: Ifa6812b2dc651fc1b19cdeeee19e760dfe483cd9
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/3942108
Commit-Queue: Steven Bingler <bingler@chromium.org>
Reviewed-by: Dave Tapuska <dtapuska@chromium.org>
Reviewed-by: Bo Liu <boliu@chromium.org>
Reviewed-by: Rakina Zata Amni <rakina@chromium.org>
Reviewed-by: Robert Sesek <rsesek@chromium.org>
Reviewed-by: Alex Moshchuk <alexmos@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1078764}
NOKEYCHECK=True
GitOrigin-RevId: a07ae73ab7f2419ed9e4634d31c4471d4b7f69b7