[inspector] Introduce a restricted instrumentation pause
We introduce a new instrumentation pause state in the renderer.
The instrumentation pause is entered on script creation
if the instrumentation breakpoints are enabled with
Debugger.setInstrumentationBreakpoint. The purpose of the pause
is to give the debugger a chance to insert its breakpoints
correctly before the script runs.

In this patch, we restrict the messages that are handled during
the instrumentation pause to the ones that are handled by the
devtools IO session (those are mostly breakpoint management
messages). Handling of other messages are postponed until the
instrumentation pause resumes.

Without this patch, the instrumentation pause behaves the same
as the normal pause, causing some undesirable interactions
with layouting, navigation, etc. (e.g., see crbug.com/1304095).

Note that the instrumentation breakpoints/pauses are
an experiment in DevTools,  currently off by default.

Bug: crbug.com/1354043
Change-Id: I44e9e8473fcb10a2868974aa8d20a07e9b00130d
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/3891025
Commit-Queue: Jaroslav Sevcik <jarin@chromium.org>
Reviewed-by: Andrey Kosyakov <caseq@chromium.org>
Reviewed-by: Daniel Cheng <dcheng@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1079116}
NOKEYCHECK=True
GitOrigin-RevId: f96df25cdfcaff34c828612347532d9e8d672d9f