Remove direct-sockets- prefix from direct sockets WPT file names
Presubmit is blocking the submission of some expectation files related
to these tests because the file name is too long for Windows. The file
path has semantic meaning so the only option I can see for reducing the
path length is renaming the tests, and the direct-socket- prefix seemed
like a good candidate as these files are in a direct-sockets/ folder.

The problematic expectation file would be in:
third_party/blink/web_tests/flag-specific/disable-site-isolation-trials
followed by the full path to the test, which combined is over the 200
character path limit we enforce for Windows.

Bug: 1333966
Change-Id: Ia26a516b4cfea40e59e58aa92415ad4aea5e9a48
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/4077065
Reviewed-by: Andrew Rayskiy <greengrape@google.com>
Commit-Queue: Andrew Rayskiy <greengrape@google.com>
Cr-Commit-Position: refs/heads/main@{#1079153}
NOKEYCHECK=True
GitOrigin-RevId: 7fb75dbb7dbcbb67755a9d9fda862b30c54a816e