Add MimeType() Interface to Image
This CL adds an MimeType() method to Image. It returns a static local
WTF::AtomicString. It is to replace the current Image::FilenameExtension() which creates a string each time which can be
avoid by this new MimeType(). Also, this new method is to return MIME
types. Note that jxl format image does not have MIME types now.

The replacement of the existing Image::FilenameExtension () would come
in a follow-up CL.

This is brought up by the discussion in a previous CL comment. See
https://chromium-review.googlesource.com/c/chromium/src/+/4027425/
comment/c06fde7f_2293b653/

Bug: 1382162
Change-Id: I5e78d1928c002de186e989d6f58049f3764a7250
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/4052107
Reviewed-by: Fredrik Söderquist <fs@opera.com>
Reviewed-by: Stephen Chenney <schenney@chromium.org>
Commit-Queue: Hao Liu <haoliuk@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1079272}
NOKEYCHECK=True
GitOrigin-RevId: b9bf2a70fb2e4492441df30fe8363ae2f2582bc8