Add //third_party/androidx_javascriptengine
The WebView team is shipping this client library that can be used by
Android apps to evaluate JavaScript and WASM.
The library is a thin interface built and shipped through AndroidX.

The actual service implementation of this library is shipped via the Android WebView APK.
In order to enable local development and end to end testing,
we mirror the JavaScriptEngine library into Chromium.

Since we require only a directory from the AndroidX repository, we
mirror a git repository created to track changes in the subdirectory
in AndroidX. This is maintained by the Git Infra team.

This library is not shipped as part of Chromium - it is only used for
local development and testing in Chromium CQ/CI.

Bug: 1393357
Change-Id: I46bfeebd81ee87e2ec1c750d099e794ea6b299fb
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/4055566
Reviewed-by: Richard Coles <torne@chromium.org>
Reviewed-by: Robert Sesek <rsesek@chromium.org>
Reviewed-by: Yaron Friedman <yfriedman@chromium.org>
Commit-Queue: Abhijith Nair <abhijithnair@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1079705}
NOKEYCHECK=True
GitOrigin-RevId: 2eed7002e3d6465a5f3e8fe0913249ccea994603