[canvas] Add a feature to conserve memory for backgrounded 2D/3D canvas
When a canvas is in a background page, we should try to reclaim some of
its memory footprint. This is the case for 2D canvas, which uses
hibernation, and for 3D canvas, where we send a
ContextVisibilityHintCHROMIUM() command. However, in both cases, we do
not ask the context to aggressively free resources, which has
consequence for (e.g.) transfer buffers.

Concretely, enabling this feature on
https://webglsamples.org/aquarium/aquarium.html saves 16MiB of transfer
memory, and up to 10s of MiB for OOP raster 2D canvas with
hibernation. It is also consistent with the intent of the code, and with
the raster context, which does set it when visibility changes.

See uploaded traces in the linked bug for details.

The feature is disabled by default, so this commit should not change
behavior.

Bug: 1355255
Change-Id: Ia0493b050f6ed991cdff46201cc6b177ecd8e7bb
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/4013426
Reviewed-by: Sunny Sachanandani <sunnyps@chromium.org>
Commit-Queue: Benoit Lize <lizeb@chromium.org>
Reviewed-by: Kenneth Russell <kbr@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1079823}
NOKEYCHECK=True
GitOrigin-RevId: 96962219634dd0657a0bac512f88e6c949f7eabc