iwa: Add --isolated-context-origins flag to enable Direct Sockets in WPT
This adds a content_shell specific --isolated-context-origins flag that
is read in ShellContentBrowserClient::IsIsolatedContextAllowedForUrl to
enable the IsolatedContext IDL attribute (which guards the Direct
Sockets API) in Web Platform Tests.

This also updates test expectations to mark the tests as failing
when site isolation is disabled, which is expected because isolation
is required for Direct Sockets.

Bug: 1333966
Change-Id: Ie2d08f2f053e9d3257b2dfb9f10525f9a23137cb
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/4068163
Reviewed-by: Andrew Rayskiy <greengrape@google.com>
Reviewed-by: danakj <danakj@chromium.org>
Reviewed-by: Weizhong Xia <weizhong@google.com>
Commit-Queue: Robbie McElrath <rmcelrath@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1080032}
NOKEYCHECK=True
GitOrigin-RevId: bb5fe2bbff72a16d2d782c8e507d62f1f058b335