LayoutNG: Fix PositionedObjects for fieldset content box
This is a follow-up of crrev.com/c/3767050.

It made ComputeIsAbsoluteContainer(old_style) not workable because the
function refers to the current style of the parent fieldset. We failed
to clear PositionedObjects() for fieldset content boxes because of it.

LayoutObject has a cached value for ComputeIsAbsoluteContainer(), and
we don't need to call ComputeIsAbsoluteContainer() for old_style now.

Bug: 1395688
Change-Id: I4d3da0c94d568a9935da01ae5fd0b86429c0ae7f
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/4082352
Auto-Submit: Kent Tamura <tkent@chromium.org>
Reviewed-by: Koji Ishii <kojii@chromium.org>
Commit-Queue: Koji Ishii <kojii@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1080181}
NOKEYCHECK=True
GitOrigin-RevId: 404d6d4b1b4b0f40818e532a5b1c8ac0b4da8c81