[Passwords] Preview password generation suggestions on hover.
This Cl allows to preview the generated password suggestion upon
mouse hover, in the same way this is done for other password
suggestions: they are previewed until the user clicks on them.

The only difference is that for normal password suggestions passwords
are shown in the masked way, but generated passwords are not masked on
preview to give the user a better understanding of what will happen
with the value that we suggest.
Mock: https://screenshot.googleplex.com/3zUSzj2vB6RgLbV

Fixed: 1359341
Change-Id: Ic4332466697c9116f0b1e3338e381fb8e27c09f2
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/3899104
Reviewed-by: Joey Arhar <jarhar@chromium.org>
Reviewed-by: Mike West <mkwst@chromium.org>
Reviewed-by: Marc Treib <treib@chromium.org>
Reviewed-by: Mohamed Amir Yosef <mamir@chromium.org>
Commit-Queue: Maria Kazinova <kazinova@google.com>
Reviewed-by: Mason Freed <masonf@chromium.org>
Reviewed-by: Christoph Schwering <schwering@google.com>
Cr-Commit-Position: refs/heads/main@{#1080230}
NOKEYCHECK=True
GitOrigin-RevId: ea2b9e57bc5b61481282b0b76e4e93d79e91c74f