Limit UPF queries to 30 days
Limits all unexpected pass finder queries that access ResultDB tables
to only look at the last 30 days of data. The tables we use are set up
to drop data after 30 days, but some initial data from when the tables
were created more than a year ago is still around. This should be
cleaned up eventually, but it's also good practice to limit the data
we're looking at in the query rather than relying on the table to
only include the data we're looking for.

Also applies to the GPU tag finder, but not to the flake finders since
those already limit the data to the most recent N days.

Bug: 1394242
Change-Id: I64b55d3b7d295d8ed9d4e0befb2cd20d9587fcd2
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/4083988
Reviewed-by: Yuly Novikov <ynovikov@chromium.org>
Commit-Queue: Yuly Novikov <ynovikov@chromium.org>
Auto-Submit: Brian Sheedy <bsheedy@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1080387}
NOKEYCHECK=True
GitOrigin-RevId: 3295822e1911e6c075d8f676b1dc903dceb0cb68