Temporarily restore cross-site prefetch no-referrer usage
To undo breakage of cross-site prefetch for sites needing a mechanism
to specify an acceptable referrer policy, we restore the originally
shipped behaviour of assuming no-referrer for this case instead of
rejecting the prefetch attempt.

This behaviour goes against the spec. We will enforce this again when we
ship SpeculationRulesReferrerPolicyKey.

For clarity, this workaround is *only* for attempts that would be
rejected. For cross-site prefetch attempts that do have an acceptable
policy, we respect that policy.

Bug: 1398772
Change-Id: I8ddcb868707cc37c28815b3f4b216c89a3929501
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/4086682
Auto-Submit: Kevin McNee <mcnee@chromium.org>
Reviewed-by: Jeremy Roman <jbroman@chromium.org>
Commit-Queue: Jeremy Roman <jbroman@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1080480}
NOKEYCHECK=True
GitOrigin-RevId: e2479c6b5612ea2444bf4942abf095b7a49b0966