[scroll-animations] Tree-scoped timeline lookups
This CL implements tree-scoped lookups for animation-timeline.

Previously, we could search for a timeline match on a given element
with a normal lookup in a hash table. This is unfortunately not enough
anymore, since tree-scoped lookups don't require exact matches, but
instead match any name in the target tree-scope or higher.
Therefore the lookup process now linearly iterates through the items
of the hash table. (We still store the timelines in a hash table
despite this linear traversal, because we use exact-match hash
lookups during CSSAnimations::CalculateTimelineUpdate).

Note: Adding a ::part() test exposed an issue with a call to
ParentOrShadowHostNode, since we'd pass a ShadowRoot to
LayoutTreeBuilderTraversal, which is not allowed. Hence this CL
changes that call to ParentOrShadowHostElement.

Fixed: 1382876
Change-Id: I4aa5ef03a663e201e97810b1f668036b90958316
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/4079252
Reviewed-by: Rune Lillesveen <futhark@chromium.org>
Commit-Queue: Anders Hartvoll Ruud <andruud@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1080561}
NOKEYCHECK=True
GitOrigin-RevId: e073524c2e259dc3e8d0ec629471b6d9017aef41