Shared Storage: Change to per-key instead of per-origin expiration.
We change to expiring keys 30 days after their most recent write,
instead of entire origins. See
https://github.com/WICG/shared-storage/issues/45 for more information.

`Length()`, `Keys()`, and `Entries()` are scoped to only the last 30 days.

If `Get()` finds an expired entry, it returns the entry with an `OperationResult::kExpired`, so that
`SharedStorageWorkletHost::SharedStorageGet()` can treat it the same
as if it received `OperationResult::kNotFound`, and so that `Set()`
and `Append()` can delete the expired entry before setting.

We moreover auto-purge the database of stale entries based on a timer,
based on the previous infra that purged stale origins, except that we
reduce the interval for the initial purge from 15 minutes to 2 minutes.

Bug: 1218540,1385197
Change-Id: Ifef109ca3bd5906d28214bd4dfded350f0c2ea8e
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/4062745
Reviewed-by: Yao Xiao <yaoxia@chromium.org>
Reviewed-by: Mike West <mkwst@chromium.org>
Commit-Queue: Cammie Smith Barnes <cammie@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1080707}
NOKEYCHECK=True
GitOrigin-RevId: df7ab6172feef1f416dabe1cde4d5685635a94f8