externs generator: require generated externs stay in sync
Whenever we update this tool, make sure the changes are reflected in
all the generated files.  This makes it easier to review changes in
here, and to make sure we don't force devs updating APIs to include
a lot of unrelated/new changes.

Then do a single refresh against all the externs so the tests pass.

Bug: 469920
Change-Id: I88dcd46f187a2c71e47cbbc77012f96cb03b7c59
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/2215902
Reviewed-by: Devlin Cronin <rdevlin.cronin@chromium.org>
Commit-Queue: Mike Frysinger <vapier@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1080726}
NOKEYCHECK=True
GitOrigin-RevId: 829eaa2da39f7529b0129f7d2622e4e5d0d92414