Update PaintImage in CopyEI2T when image uploading fallback to CPU path
This is a bug in CopyEI2T when fallback has been triaged. In the fallback path, image backed by GPU is MakeUnaccelerated() first.
In that case, paint_image which created from origin image should be regenerated but it doesn't.

This CL fix the issue.

Bug: chromium:1309194
Change-Id: I3054de93694a3a1b7a925c993c368bf9a19c1b53
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/4069320
Reviewed-by: Corentin Wallez <cwallez@chromium.org>
Reviewed-by: Austin Eng <enga@chromium.org>
Commit-Queue: Corentin Wallez <cwallez@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1080869}
NOKEYCHECK=True
GitOrigin-RevId: c52a915593debb8017cb0a2e96f1e62b2190c3c9