Fix LCP start time logic for non TAO images
This is to fix the logic of LCP start time for Non-Tao images and also
clean up of the previous CL https://crrev.com/c/4027513. The previous CL only accounts for the first non Tao image as LCP candidate.

The wpt in previous CL is flaky on Mac11-arm64. That is because somehow
FCP and LCP are set not by the same updates. We would expect so for a
page with an image and nothing else.

That wpt was re-written as non-tao-image-load-before-fcp-render-at-fcp.sub.html. That failure is no longer relevant as the new logic compares FCP with LCP.loadTime. We simply need to verify the logic is correct in the wpt. That is, if LCP.loadTime < FCP, LCP startTime should be the same as FCP, otherwise LCP startTime should be the same as LCP loadTime.

Also start time is passed as an additional arg to WindowPerformance::OnLargestContentfulPaintUpdated(), instead of being derived. This makes the logic easier to follow.

Bug: 1345948
Change-Id: I4451c8c46dc9c00eb5ecd2da743210d3a94004bf
Cq-Include-Trybots: luci.chromium.try:mac11-arm64-rel,mac12-arm64-rel
Include-Ci-Only-Tests: true
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/4057886
Reviewed-by: Yoav Weiss <yoavweiss@chromium.org>
Commit-Queue: Hao Liu <haoliuk@chromium.org>
Reviewed-by: Dominic Farolino <dom@chromium.org>
Reviewed-by: Bo Liu <boliu@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1080910}
NOKEYCHECK=True
GitOrigin-RevId: 9785a24233f18e2384aca2d0c15b663610803e28