Error message when WebGPU context provider fails
Previously when the WebGPUGraphicsContext3DProvider failed to create it
would simply resolve the promise with `null`, which led to failures with
no apparent cause. This change make use of the regular adapter creation
callback so that an error message can be reported in the console for
debugging purposes.

Bug: 973017
Change-Id: I86488a1cd692f82a0017910d6b08691d69395ad6
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/4089524
Commit-Queue: Brandon Jones <bajones@chromium.org>
Reviewed-by: Kai Ninomiya <kainino@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1081213}
NOKEYCHECK=True
GitOrigin-RevId: 16501b1f37656732d154a19249e9ebdf8f740275