[selectmenu] Add UA style sheet to position the listbox
This patch removes all the proprietary positioning code for selectmenu
listbox popover, and replaces it with a UA style sheet. This also
makes anchor positioning a dependency for selectmenu.

There's a caveat that we don't have a way in UA style sheet to ensure
selecting an externally provided listbox [1]. This patch works around
it with `selectmenu [behavior=listbox]` and also leaves a TODO.

[1] https://github.com/openui/open-ui/issues/645

Bug: 1197720
Change-Id: I6308089e4be858432969b87eabd27b9e0ea6c435
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/4082046
Commit-Queue: Xiaocheng Hu <xiaochengh@chromium.org>
Reviewed-by: Anders Hartvoll Ruud <andruud@chromium.org>
Reviewed-by: Mason Freed <masonf@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1081249}
NOKEYCHECK=True
GitOrigin-RevId: 7b60f9a0bbaa5536887a1e23689d5cef52f0a48c