Prerender2 CSP: Implement script-src 'inline-speculation-rules'
This patch implements CSP 'inline-speculation-rules' source for
<script type="speculationrules">. This source is available only for
script-src directive, and can permit inline script tags that has
the type attribute with "speculation-rules". This allows developers
to permit inline script only for speculationrules.

With this patch, the change is applied only for isolated worlds
where we run Chrome Extensions. But it will be extended to the
open web platform once we publish a patch spec and follow the I2S.

Bug: 1382361
Change-Id: I69b1f3e61e64c3f956964fe59e2e847239654c81
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/4056835
Reviewed-by: Dominic Farolino <dom@chromium.org>
Commit-Queue: Takashi Toyoshima <toyoshim@chromium.org>
Reviewed-by: John Abd-El-Malek <jam@chromium.org>
Reviewed-by: Mike West <mkwst@chromium.org>
Reviewed-by: Hiroki Nakagawa <nhiroki@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1081341}
NOKEYCHECK=True
GitOrigin-RevId: 9a89f759b68a49bb969ec1a7a4d3d966cb5707bc