[container-queries] Start CQ search for :host:part() correctly
We are supposed to start the search for container queries at the
ultimate originating element, but we did not respect this for
:host:part() rules. This is because the first iteration of a for-loop
in MatchPseudoPartRules starts with the part_matching_element, which
is used as the starting element for the container queries search.

To meet the "ultimate originating element" criteria, we should instead
start at that element's host.

Fixed: 1399539
Change-Id: Icfc4265f5dae083c15a204f78962e8a356312cdc
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/4088568
Reviewed-by: Rune Lillesveen <futhark@chromium.org>
Commit-Queue: Anders Hartvoll Ruud <andruud@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1081384}
NOKEYCHECK=True
GitOrigin-RevId: a35d5d1dd68bdb4b78dc74de51182f5ac4dda995