VT: Ensure all layers that recalculate their draw props push props too
This patch is a more comprehensive fix for the referenced bug.
Previously, we've only addressed opacity animations, but the same bug
is present for transform animations as well. The problem with
transforms is that to detect changes in the matrix, we need to do
a lot of float comparisons. It seems better to just unconditionally
say that if we recalculated draw properties, then that layer needs
to push properties as well.

Note for picture layer impls, this is still a no-op, since they
always push their properties.

R=pdr@chromium.org

Bug: 1385432
Change-Id: I6df15e6e344507f26336698e4bf890728690da52
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/4081632
Reviewed-by: Philip Rogers <pdr@chromium.org>
Auto-Submit: Vladimir Levin <vmpstr@chromium.org>
Commit-Queue: Vladimir Levin <vmpstr@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1081505}
NOKEYCHECK=True
GitOrigin-RevId: c7889d3da725ed2f6b6efd7dde8ff9e27c1debe9