Link ServiceWorkerVersion in-flight requests to event queue idle timeout
ServiceWorkerVersion in-flight requests can block idle timeout request
from event queue. However they couldn't extend the idle timer. As a
result extension service worker events, which leverage in-flight
requests, cannot guarantee 30 seconds idle time.

In this CL we reuse the idle timeout callback logics in
ServiceWorkerEventQueue. That is, 1) cancel the idle timeout when
starting an event; 2) check event queue when finished an event, schedule
the idle timeout if the event queue is empty.

Before checking event queue we also check if there is any in-flight
requests. This makes sure idle timeout is scheduled only when in-flight
request queue is empty too.

ServiceWorkerEventQueue logic remain unchanged. There might be a case
that idle timeout is scheduled when there are ongoing in-flight
requests.
Which is okay because ongoing in-flight requests will block termination
request. And by the time in-flight request finishes the idle timeout may
be rescheduled.

R=lazyboy@chromium.org, rdevlin.cronin@chromium.org

Change-Id: I901727d70d1130d0c55fe38f38e507fdc91c4304
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/4022172
Reviewed-by: Hiroki Nakagawa <nhiroki@chromium.org>
Commit-Queue: Richard Zhang <richardzh@chromium.org>
Reviewed-by: Kouhei Ueno <kouhei@chromium.org>
Reviewed-by: Yoshisato Yanagisawa <yyanagisawa@chromium.org>
Reviewed-by: Daniel Cheng <dcheng@chromium.org>
Reviewed-by: Devlin Cronin <rdevlin.cronin@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1081592}
NOKEYCHECK=True
GitOrigin-RevId: 9151f87dfbf2f92e5b5549b4eca88a67b70bd1a5