Fix a crash when a popoverhide event handler changes the popover type
This code would cause a crash, previously:

  popover.showPopover();
  popover.addEventListener('popoverhide',() => {
    popover.setAttribute('popover',"manual");
  });
  popover.hidePopover();

The problem was that hidePopover first removes the "auto" popover from
the stack, then fires the popoverhide event, and that event handler
then can't find the popover in the stack. There was already a check
for this, but it was roundabout. Now, just check that the stack
doesn't contain this popover, and early-out.

Bug: 1307772
Change-Id: I3ba19347cf987013e28bf67b1d6707794fed4398
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/4074665
Commit-Queue: Mason Freed <masonf@chromium.org>
Reviewed-by: Joey Arhar <jarhar@chromium.org>
Commit-Queue: Joey Arhar <jarhar@chromium.org>
Auto-Submit: Mason Freed <masonf@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1081630}
NOKEYCHECK=True
GitOrigin-RevId: 732af091a1d4f08f960c6e07db32458fd958a0df