Invalidate <filter> while parsing if it has an associated resource
If a paint takes place while the HTML parser is in the middle of parsing
a <filter> element, the paint would end up with a "partial" filter chain.
Since the ChildrenChange::ByParser() flag is used to early out, any
filter primitive elements that had not yet been "seen" would not end up
triggering invalidation, breaking the expected filter rendering.

If an SVGFilterElement has an associated SVGResource, it means that
there exists (or at least has existed) a client element that will be
using the filter. Use the presence of an associated SVGResource as a
signal that invalidation should take place during parsing as well.

Bug: 1368369
Change-Id: Ieb05ecf040051f47774113e91e604ff22ed6c2a4
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/4097086
Commit-Queue: Fredrik Söderquist <fs@opera.com>
Reviewed-by: Xianzhu Wang <wangxianzhu@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1081885}
NOKEYCHECK=True
GitOrigin-RevId: 79d93862a96600b0a9f8f35a804f7aa256682e9b