blink: Modifies CalculateSelectionBounds to dispose Range
As this range is only needed in calculating the bounds there is no
need for it to stay attached to the Document. Keeping the Range
attached to the document results in updating the Range until a gc,
which is unnecessary.

This patch is part of a bigger set of patches. Combined patches are
here: https://chromium-review.googlesource.com/c/chromium/src/+/4082153

Change-Id: Ibac5beea0c1255d994684899bef5641dd21111cb
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/4097115
Commit-Queue: Scott Violet <sky@chromium.org>
Reviewed-by: Xiaocheng Hu <xiaochengh@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1082106}
NOKEYCHECK=True
GitOrigin-RevId: dadf0e4a29c66d6e64da5aeb3c5f7e719f641b26