blink: adds flag for pseudo element to rare data
Whether a node is a pseudo element is needed when traversing
the dom:

https://chromium.googlesource.com/chromium/src/+/fa6cb47ea509fcaf4d459cd63e5b6a0270561d32/third_party/blink/renderer/core/dom/flat_tree_traversal.cc#153

This code path is hit extensively and currently involves a virtual
call. As pseudo elements are not that common, this adds a flag to
rare-data to determine that is used instead.

By itself this patch doesn't move the metrics much, but part of a
bigger set of patches it does.

This patch is part of a bigger set of patches. Combined patches are
here: https://chromium-review.googlesource.com/c/chromium/src/+/4082153

Change-Id: Icfe3528745d5e0f8daee370174f8f1cd0f14380a
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/4086823
Reviewed-by: Xiaocheng Hu <xiaochengh@chromium.org>
Commit-Queue: Scott Violet <sky@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1082107}
NOKEYCHECK=True
GitOrigin-RevId: 6e7c2ce0738f4320db07a904872483d9808aae08