WebNN: Implement the build method of sigmoid operation
This CL implements the sigmoid() methods of MLGraphBuilder. The sigmoid
can be built and connected with input and output operands. Or it can be
built as a standalone operator that represents an activation function
fused into another operation, such as conv2d. This CL also implements
the unit tests for both scenarios.

Bug: 1273291
Change-Id: Idd09cdf97b40a04f7427a50304992354db2914f8
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/4081237
Reviewed-by: Jiewei Qian <qjw@chromium.org>
Commit-Queue: Jiewei Qian <qjw@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1082126}
NOKEYCHECK=True
GitOrigin-RevId: 7614c6ed0a2c8fdcdd56564c41fb1982c360a097