Fenced frames: add urn to config/redacted config object
At the moment, the web platform accessible inner fenced frame config
object is a shell that can't be used to navigate a fenced frame. This is
because, right now, all fields are redacted and there is no way for an
inner config to determine what navigation it is associated with.

On the browser side, there exists a URN to fenced frame config mapping,
`FencedFrameURLMapping::urn_uuid_to_url_map_`. When a fenced frame
navigates to a URN, the browser pulls the URL from the associated config
object in the mapping and then navigates the fenced frame to that URL.
Without access to that URN in the inner config object, the web platform
cannot tell the browser what to navigate a fenced frame to.

This CL introduces a field, `urn_`, to the fenced frame config object.
That value is sent from the browser to the renderer.

In a follow up CL, logic will be added to send the urn over mojo to
the renderer and read it into the web platform accessible
`FencedFrameConfig` object.

Change-Id: Ibb708256099a46587800124524eb2d44a157bfe7
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/4090613
Reviewed-by: Garrett Tanzer <gtanzer@chromium.org>
Reviewed-by: Dominic Farolino <dom@chromium.org>
Commit-Queue: Liam Brady <lbrady@google.com>
Cr-Commit-Position: refs/heads/main@{#1082128}
NOKEYCHECK=True
GitOrigin-RevId: 80288c566b05e7054f4f26033e75d5dae04d3b20