blink: Selection related optimizations
This consists of the following:
. Use a const& return value from selection where possible.
. Change TextControlElement to have a single function responsible for
  calculating complete selection.

Doing the latter allows selection computation to use the value
calculated for the start as the end if the start and end are the same.

This patch is part of a bigger set of patches. Combined patches are
here: https://chromium-review.googlesource.com/c/chromium/src/+/4082153

Change-Id: I451a2d383c52d5bebd478449aadfe2d772720a61
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/4087045
Reviewed-by: Xiaocheng Hu <xiaochengh@chromium.org>
Commit-Queue: Scott Violet <sky@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1082202}
NOKEYCHECK=True
GitOrigin-RevId: 83fdec5d3205d055690f67a15de6f3b380d9f5ed