Give default action to date and time picker buttons.
Certain complex HTML elements have components that are expected to be
clickable, but they don't have an attached listener and the native role
doesn't match. This is the case of the picker button in
<input type=date|time>.

Luckily, we label these components with a ShadowPseudoId we use for
our own purposes, so we can use them in accessibility code to assign
the correct action.

We add a test checking the action is exposed, to prevent regressions;
to implement the test we had to expose a new method HasDefaultAction()
to web tests through the WebAXObjectProxy object.

Bug: 1379347
Change-Id: I4511b7d17d67db465cbb5dd6f7525e9ba8a2ccd6
AX-relnotes: give default action to date and time picker buttons.
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/4020585
Reviewed-by: Frédéric Wang <fwang@igalia.com>
Reviewed-by: Aaron Leventhal <aleventhal@chromium.org>
Commit-Queue: Jacobo Aragunde Pérez <jaragunde@igalia.com>
Cr-Commit-Position: refs/heads/main@{#1082400}
NOKEYCHECK=True
GitOrigin-RevId: bc6c8b98ed3971011c67e73242ffe17380a1a8f2