Reland "Fix the popover-attribute-basic.tentative.html test"
This is a reland of commit ae325b42a8d837badc8a10038664c0865ee8e599

The failures were timeouts, and it is a long test, so I've added
a meta-timeout=long to see if that fixes the flakiness. Patchset 1 is the original patch, for comparison.

********
SHERIFFS: please don't revert this CL if the flakiness in the
popover-attribute-basic.tentative.html test continues. This CL
contains some fixes for that test, and those are separate from the
flakiness. Just mark it flaky and assign me a bug and I'll take
a look. Thanks!
********

Original change's description:
> Fix the popover-attribute-basic.tentative.html test
>
> This test has been disabled for some time, while [1] was landed. Then
> [2] landed, which changed the event names/behavior, but that CL forgot
> to update this test. Additionally, there was a crash that was fixed
> in [3]. This CL cleans it up and re-enables it finally.
>
> [1] https://chromium-review.googlesource.com/c/chromium/src/+/4074305
> [2] https://chromium-review.googlesource.com/c/chromium/src/+/4035832
> [3] https://chromium-review.googlesource.com/c/chromium/src/+/4074665
>
> Bug: 1307772
> Change-Id: I3f0b9fb604a04cb0cc28370f40d8b02d5fce11ac
> Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/4090644
> Commit-Queue: Mason Freed <masonf@chromium.org>
> Auto-Submit: Mason Freed <masonf@chromium.org>
> Reviewed-by: Joey Arhar <jarhar@chromium.org>
> Cr-Commit-Position: refs/heads/main@{#1081664}

Bug: 1307772
Fixed: 1400166
Change-Id: I5bdf76970caf3432f3a9b6a84660548d8914fa8d
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/4099315
Reviewed-by: Joey Arhar <jarhar@chromium.org>
Commit-Queue: Joey Arhar <jarhar@chromium.org>
Auto-Submit: Mason Freed <masonf@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1082570}
NOKEYCHECK=True
GitOrigin-RevId: 227cd371715d2c01c0353ae19c3b282ad3dfe9ea