Add UMA metric for blink main thread task queueing time
Queueing time metrics were previously added in:

https://chromium-review.googlesource.com/c/chromium/src/+/1606191

... but were later removed in an effort to reduce the cost of per-task
metrics collection. This patch uses a different methodology than the
previous implementation; in particular:

- The previous metric was collated by queue type; this CL collates by
queue priority.

- The previous metric recorded aggregate queueing time over all tasks
for a given queue type; this CL records a sample per task.

- The previous metric used TaskTiming.queue_time as the first
timestamp; this CL uses Task::GetDesiredExecutionTime(). That means
that for timer tasks we include the time interval between when the
timer theoretically expires, and when the event loop notices that the
timer has expired and puts the task on the active queue.

Bug: 1384353
Change-Id: I85581867bceeb31ecac5ea179020f1a382f96993
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/4042361
Reviewed-by: Scott Haseley <shaseley@chromium.org>
Commit-Queue: Stefan Zager <szager@chromium.org>
Reviewed-by: Stephen Chenney <schenney@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1082654}
NOKEYCHECK=True
GitOrigin-RevId: aba7c8ddd2b6420e3f172479cfeb83fbeeded38f