Add shared worker IsSecureContext tests to 3PSP test suite
It's possible for two contexts to be same-origin but not both
be considered secure contexts, so for shared workers the spec
indicates that an error should be raised if these two contexts
would otherwise share a shared worker. When storage partitioning
is enabled, though, this case is prevented by the ancestor chain
bit in the storage key. When an insecure ancestor exists in a
sequence of otherwise secure, same-site frames, the ancestor
chain bit in the shared worker will be set to kCrossSite and
will be partitioned separately from a shared worker created from
a sequence of secure, same-site frames. Thus, an insecure
context can no longer share a shared worker with a secure
context.

This CL pulls two shared worker tests for these cases into the
third-party-storage-partitioning virtual test suite so that we
can document this case (and when we enable storage partition
as part of experimental web platform features, we will still
want to test that this part of the spec is upheld when SP is
disabled).

Bug: 1354245
Change-Id: Ife9ab3cd9f6cb897f14b3a3b0ff251244262d127
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/4105684
Reviewed-by: Dominic Farolino <dom@chromium.org>
Commit-Queue: Andrew Williams <awillia@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1083129}
NOKEYCHECK=True
GitOrigin-RevId: 8c7035083035377298a8d355ea5846a506003a50